+++ app/mcp/v1/tools/duckdb_reader.py
@@ -2,7 +2,7 @@ from typing import Any
 from fastmcp import FastMCP
 
 from app.schemas.record import RecordType, IntervalType, HealthRecordSearchParams
-from app.services.health.duck_queries import (
+from app.services.health.duckdb_queries import (
     get_health_summary_from_duckdb,
     search_health_records_from_duckdb,
     get_statistics_by_type_from_duckdb,
