+++ app/services/duckdb_client.py
